afk auto-clears when user sends a message + antilink ignores ppl with manage server

+++ bot.py
@@ -143,36 +143,43 @@ async def on_message(message):
         pass
     else:
         return
+    data = await bot.db.afk.find_one({"id": message.author.id})
+    data = await bot.db.afk.update_one({"id": x.id}, {"$set": {"status": False}})
+    if data.get("status", None):
+        await message.channel.send(f"Oh hey {message.author.mention}, welcome back! For your convenience I cleared your AFK status.")
     if re.match(f"^<@!?{bot.user.id}>$", message.content):
         await message.channel.send(f"{bot.get_emoji(430853515217469451)} BAH! Why you :regional_indicator_p:ing me? Anyway, I'm dat banana bot, so nice to meet you. I do a LOT of kewl stuff, like music, starboard, welcome/leave messages, Canvas, and so much more! All it takes is `*help` to see the powers I got! {bot.get_emoji(430853629570711562)}")
     if re.findall(r"(http(s)://|)(discord\.gg|discord\.io|discordapp\.com/invite)\S+", message.content):
-        x = await bot.db.antilink.find_one({"id": message.guild.id})
-        if not x:
-            return
-        if not x['status']:
-            return
-        try:
-            await message.delete()
-        except discord.Forbidden:
+        if message.author.guild_permissions.manage_guild:
             pass
-        await message.channel.send(f"Hey, {message.author.mention}! No advertising allowed in this server. Get that invite link out of here!")
-        if await modlog_check(message.guild.id):
+        else:           
+            x = await bot.db.antilink.find_one({"id": message.guild.id})
+            if not x:
+                return
+            if not x['status']:
+                return
             try:
-                lol = bot.get_channel(await get_modlog_channel(message.guild.id))
-                em = discord.Embed(color=0xf9e236, title="Invite Posted")
-                em.description = textwrap.dedent(f"""
-                {bot.get_emoji(430340802879946773)} Sent by **{str(message.author)}**
-
-                :hash: In channel {message.channel.mention}
-
-                :link: Link:
-                {message.content}
-
-                """)
-                em.timestamp = message.created_at
-                await lol.send(embed=em)
-            except KeyError:
+                await message.delete()
+            except discord.Forbidden:
                 pass
+            await message.channel.send(f"Hey, {message.author.mention}! No advertising allowed in this server. Get that invite link out of here!")
+            if await modlog_check(message.guild.id):
+                try:
+                    lol = bot.get_channel(await get_modlog_channel(message.guild.id))
+                    em = discord.Embed(color=0xf9e236, title="Invite Posted")
+                    em.description = textwrap.dedent(f"""
+                    {bot.get_emoji(430340802879946773)} Sent by **{str(message.author)}**
+
+                    :hash: In channel {message.channel.mention}
+
+                    :link: Link:
+                    {message.content}
+
+                    """)
+                    em.timestamp = message.created_at
+                    await lol.send(embed=em)
+                except KeyError:
+                    pass
     # levelup = await bot.db.level.find_one({"id": message.guild.id})
     # if levelup:
     #     try:
